+++ src/main/java/com/team45/service/ChatService.java
@@ -6,14 +6,14 @@ import com.team45.entity.ChatRoom;
 import java.util.List;
 
 public interface ChatService {
-    public List<ChatRoom> chatRoomProductList(int pno);
-    public ChatRoom chatRoomGetNo(int roomNo);
-    public ChatRoom chatRoomInsert(String memId, int pno);
-    public int chatRoomBlockUpdate(int roomNo);
+    public List<ChatRoom> chatRoomProductList(Long pno);
+    public ChatRoom chatRoomGetNo(Long roomNo);
+    public ChatRoom chatRoomInsert(String memId, Long pno);
+    public int chatRoomBlockUpdate(Long roomNo);
 
-    public List<ChatMessage> chatMessageList(int roomNo);
+    public List<ChatMessage> chatMessageList(Long roomNo);
     public ChatMessage chatMessageInsert(ChatMessage chatMessage);
-    public int chatMessageReadUpdate(int chatNo, String sender);
-    public int chatMessageReadUpdates(int roomNo, String sender);
-    public int chatMessageRemoveUpdate(int chatNo);
+    public int chatMessageReadUpdate(Long chatNo, String sender);
+    public int chatMessageReadUpdates(Long roomNo, String sender);
+    public int chatMessageRemoveUpdate(Long chatNo);
 }
